U'd version parsing tests.

+++ DbSchema.Version/Feature/Version/Type/t_version.sql
@@ -1,1 +1,1 @@
-﻿create type [schema_version].[t_version] from integer not null;
+﻿create type [schema_version].[t_version] from decimal(38,0) not null;
